+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## TODO
 - Clean the lints!
 
+## 2026-03-13
+- Fix Caiso bucket parsing
+
 ## 2026-02-25
 - Add a tbN function to calculate the max battery storage value.